// src/webparts/crudTest/interfaces/ProductInterfaces.ts
export interface IProductRes{
    Id: number;
    Title: string;
    Description: string;
    Quantity: number;
    Category: string;
    DateOfUpdata: string;
  }
  
export interface IProduct{
    id: number;
    productName: string;
    description: string;
    quantity: number;
    category: string;
    dateOfUpdate: string;
  }